refactor(ci): rebalance integration buckets by extracting asset & arkcash e2e tests
The integration matrix bucketed by file, but ark.test.ts was a 2400-line
monolith bundling unrelated concerns. Asset issuance/burn/send tests lived
in the "ark-core" bucket while the "arkade-assets" bucket held arkade-script
and liquidation instead — the names didn't match the contents, and the
indivisible ark.test.ts dominated wall-clock.

Extract the "Asset integration tests" and "ArkCash" describe blocks into
test/e2e/asset.test.ts and test/e2e/arkcash.test.ts, then regroup:
- arkade-assets: asset + arkcash + liquidation (genuinely asset-related)
- settlement-delegation: gains arkade-script
- ark-core: just ark.test.ts

Buckets go from 37/32/17/7 to 24/32/23/19 tests. ArkCash tests are few and
light, so they fold into arkade-assets rather than paying a fifth regtest
spin-up.
